requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
WTForms==3.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
email-validator==2.3.0
